requirements.txt
google-adk
pysqlite3-binary
toolbox-core
toolbox-langchain==0.1.0
pdfplumber
google-cloud-aiplatform
cloudpickle==3.1.1
pydantic==2.10.6
pytest
overrides
scikit-learn
reportlab
google-cloud-aiplatform[adk,agent_engines]
google-auth
google-cloud-storage
mcp
deprecated
